% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_renyi.R
\name{bootstrap_renyi}
\alias{bootstrap_renyi}
\title{Function for bootstrapping R?nyi transfer entropy under H0 of independence
between time series x and y.}
\usage{
bootstrap_renyi(x, lx, y, ly, q, burn = 50, constx = 0, consty = 0,
  nreps = 2, shuffles = 6, cl = NULL)
}
\arguments{
\item{x}{a vector of values}

\item{lx}{Markov order of x, defaults to 1}

\item{y}{a vector of values}

\item{ly}{Markov order of y, defaults to 1}

\item{q}{weighting parameter in Renyi transfer entropy, defaults to 0.5}

\item{burn}{number of observations that are dropped from the beginning of
the bootstrapped Markov chain}

\item{constx}{constant value substracted from transfer entropy measure x,
default to NULL (no const)}

\item{consty}{constant value substracted from transfer entropy measure y,
default to NULL (no const)}

\item{nreps}{number of replications for each shuffle}

\item{shuffles}{number of shuffles}

\item{cl}{a numeric value (defaults to number of cores - 1),
or a cluster as created by \code{\link[parallel]{makeCluster}}
that can be used by \code{\link[pbapply]{pbapply}}}
}
\value{
returns a vector
}
\description{
Function for bootstrapping R?nyi transfer entropy under H0 of independence
between time series x and y.
}
\examples{

}
\keyword{internal}
