% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_gen.R
\name{cluster_gen}
\alias{cluster_gen}
\title{Function that generates clusters and calculates frequencies for each cluster.}
\usage{
cluster_gen(x, y = NULL, lx, ly = NULL, prog = TRUE)
}
\arguments{
\item{x}{a vector of coded values}

\item{y}{a vector of coded values}

\item{lx}{x(k)}

\item{ly}{y(j)}

\item{prog}{if TRUE, x(k+1)}
}
\value{
returns a list with clusters and associated frequencies
}
\description{
Function that generates clusters and calculates frequencies for each cluster.
}
\examples{

}
