% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_sample.R
\name{code_sample}
\alias{code_sample}
\title{Codes the sample by assigning data to bins that are, for example, based on 
the quantiles of the empirical distribution of the sample.}
\usage{
code_sample(Data, Type = "quantiles", Quantile = c(5, 95), Bins = NULL,
  Limits = NULL, Scale = 1e+10)
}
\arguments{
\item{Data}{Column of a data table}

\item{Type}{Assign a certain number of bins, limits or use quantiles of 
empirical distribution to tansform data into discrete form}

\item{Quantile}{Define the quantiles to use for discretization}

\item{Bins}{Define the number of bins with equal width used for 
discretization}

\item{Limits}{Define limits used for discretization explicitly}

\item{Scale}{}
}
\value{
Function returns a data table column
}
\description{
Codes the sample by assigning data to bins that are, for example, based on 
the quantiles of the empirical distribution of the sample.
}
\examples{

}
