% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{coefs}
\alias{coefs}
\title{Extract the Coefficient Matrix from a TEResult}
\usage{
coefs(x)
}
\arguments{
\item{x}{a TEResult}
}
\value{
a Matrix containing the coefficients
}
\description{
Extract the Coefficient Matrix from a TEResult
}
\examples{
set.seed(1234567890)
n <- 1000
x <- rep(0, n + 1)
y <- rep(0, n + 1)

for (i in seq(n)) {
  x[i + 1] <- 0.2 * x[i] + rnorm(1, 0, 2)
  y[i + 1] <- x[i] + rnorm(1, 0, 2)
}

x <- x[-1]
y <- y[-1]

te_result <- transfer_entropy(x, y)
coefs(te_result)
}
