% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_sample.R, R/combine_sample_fix.R
\name{combine_sample}
\alias{combine_sample}
\alias{combine_sample}
\title{Function that combines the given time series.}
\usage{
combine_sample(x, y)

combine_sample(x, y)
}
\arguments{
\item{x}{a vector of coded values}

\item{y}{a vector of coded values}

\item{x}{a vector of coded values}

\item{y}{a vector of coded values}
}
\value{
returns a list

returns a list
}
\description{
Function that combines the given time series.

Function that combines the given time series.
}
\examples{


}
