% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode_sample.R
\name{decode_sample}
\alias{decode_sample}
\title{Function to decode a combined and coded sample of two time series.}
\usage{
decode_sample(x, coded)
}
\arguments{
\item{x}{coded time series}

\item{coded}{table of codes and associated values}
}
\value{
returns a matrix
}
\description{
Function to decode a combined and coded sample of two time series.
}
\keyword{internal}
