% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEclass.R
\name{is.TEResult}
\alias{is.TEResult}
\title{Checks if an object is a TEResult}
\usage{
is.TEResult(x)
}
\arguments{
\item{x}{an object}
}
\value{
a boolean value if x is a TEResult
}
\description{
Checks if an object is a TEResult
}
\examples{
# see ?transfer_entropy
}
