% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/te_class.R
\name{is.transfer_entropy}
\alias{is.transfer_entropy}
\title{Checks if an object is a transfer_entropy}
\usage{
is.transfer_entropy(x)
}
\arguments{
\item{x}{an object}
}
\value{
a boolean value if x is a transfer_entropy
}
\description{
Checks if an object is a transfer_entropy
}
\examples{
# see ?transfer_entropy
}
