% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_boot_step.R
\name{markov_boot_step}
\alias{markov_boot_step}
\title{Function for step-wise bootstraping the Markov chain.}
\usage{
markov_boot_step(x, lx, burn = 50)
}
\arguments{
\item{x}{a vector of coded values}

\item{lx}{Markov order of x}

\item{burn}{number of observations that are dropped from the beginning of
the bootstrapped time series}
}
\value{
returns a vector of characters
}
\description{
Function for step-wise bootstraping the Markov chain.
}
\examples{

}
\keyword{internal}
