% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEclass.R
\name{print.TEResult}
\alias{print.TEResult}
\title{Prints a transfer-entropy result}
\usage{
\method{print}{TEResult}(x, ...)
}
\arguments{
\item{x}{a TEResult}

\item{...}{}
}
\value{
invisible the text
}
\description{
Prints a transfer-entropy result
}
\examples{
# see ?transfer_entropy()
}
