% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Renyi_transfer_entropy.R
\name{Renyi_transfer_entropy}
\alias{Renyi_transfer_entropy}
\title{Function to implement Renyi transfer entropy.}
\usage{
Renyi_transfer_entropy(x, lx, y, ly, q, shuffle = TRUE, const = FALSE,
  constx = 0, consty = 0, nreps = 2, shuffles = 6,
  ncores = parallel::detectCores() - 1, type = "quantiles",
  quantiles = c(5, 95), bins = NULL, limits = NULL, nboot, burn = 50)
}
\arguments{
\item{x}{a vector of coded values}

\item{lx}{Markov order of x}

\item{y}{a vector of coded values}

\item{ly}{Markov order of y}

\item{q}{weighting parameter in Renyi transfer entropy}

\item{shuffle}{if TRUE, shuffled transfer entropy is calculated}

\item{const}{if TRUE, then shuffle is constant for all bootstraps}

\item{constx}{constant value substracted from transfer entropy measure}

\item{consty}{constant value substracted from transfer entropy measure}

\item{nreps}{number of replications for each shuffle}

\item{shuffles}{number of shuffles}

\item{ncores}{number of cores in parallel computation}

\item{type}{bins, limits or quantiles of empirical distribution to discretize
the data}

\item{quantiles}{quantiles to use for discretization}

\item{bins}{the number of bins with equal width used for discretization}

\item{limits}{limits used for discretization}

\item{nboot}{number of bootstrap replications}

\item{burn}{number of observations that are dropped from the beginning of
the bootstrapped Markov chain}

\item{parcalc}{"yes" if parallel computations for the Markov chain bootstrap}

\item{boots}{number of bootstrap samples}
}
\value{
returns a list
}
\description{
Function to implement Renyi transfer entropy.
}
\examples{

}
