% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_quiet.R
\name{set_quiet}
\alias{set_quiet}
\title{Set the quiet-parameter for all RTransferEntropy Calls}
\usage{
set_quiet(quiet)
}
\arguments{
\item{quiet}{if FALSE, the functions will give feedback on the progress}
}
\value{
nothing
}
\description{
Set the quiet-parameter for all RTransferEntropy Calls
}
\examples{
# see ?transfer_entropy
}
