% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shannon_transfer_entropy.R
\name{Shannon_transfer_entopy}
\alias{Shannon_transfer_entopy}
\title{Function to implement Shannon transfer entropy.}
\usage{
Shannon_transfer_entopy(x, y, lx, ly, nrep = 10, shuffles = 6,
  const = FALSE, shuffle = TRUE, bins, quantiles, nboot)
}
\arguments{
\item{nboot}{}
}
\value{
returns a list
}
\description{
Function to implement Shannon transfer entropy.
}
\examples{

}
