% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffle_renyi.R
\name{shuffle_renyi}
\alias{shuffle_renyi}
\title{Function to calculate the effective transfer entropy, as given by the
difference between the Renyi transfer entropy calculated from a sample and
the respective shuffled transfer entropy.}
\usage{
shuffle_renyi(x, lx, y, ly, q, nreps = 2, shuffles = 6, diff = TRUE,
  cl = NULL)
}
\arguments{
\item{x}{a vector of values}

\item{lx}{Markov order of x, defaults to 1}

\item{y}{a vector of values}

\item{ly}{Markov order of y, defaults to 1}

\item{q}{weighting parameter in Renyi transfer entropy, defaults to 0.5}

\item{nreps}{number of replications for each shuffle}

\item{shuffles}{number of shuffles}

\item{cl}{a numeric value (defaults to number of cores - 1),
or a cluster as created by \code{\link[parallel]{makeCluster}}
that can be used by \code{\link[pbapply]{pbapply}}}
}
\value{
returns a numeric scalar
}
\description{
Function to calculate the effective transfer entropy, as given by the
difference between the Renyi transfer entropy calculated from a sample and
the respective shuffled transfer entropy.
}
\examples{

}
\keyword{internal}
