% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffled_transfer_entropy.R
\name{shuffled_transfer_entropy}
\alias{shuffled_transfer_entropy}
\title{Function to calculate the effective transfer entropy, as given by the
difference between the transfer entropy calculated from a sample and the
respective shuffled transfer entropy.}
\usage{
shuffled_transfer_entropy(nrep = 10, shuffles = 6, diff = TRUE, x, lx, y,
  ly, ncores = parallel::detectCores() - 1)
}
\arguments{
\item{diff}{if TRUE, the effective transfer entropy is calculated, otherwise
only the shuffled transfer entropy is returned}

\item{x}{a vector of coded values}

\item{lx}{x(k)}

\item{y}{a vector of coded values}

\item{ly}{y(j)}

\item{ncores}{the number of cores}
}
\value{
returns a numeric scalar
}
\description{
Function to calculate the effective transfer entropy, as given by the
difference between the transfer entropy calculated from a sample and the
respective shuffled transfer entropy.
}
\examples{

}
