% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEclass.R
\name{summary.TEResult}
\alias{summary.TEResult}
\title{Prints a summary of a transfer-entropy result}
\usage{
\method{summary}{TEResult}(x, ...)
}
\arguments{
\item{x}{a TEResult}

\item{...}{additional arguments}
}
\value{
invisible the text
}
\description{
Prints a summary of a transfer-entropy result
}
\examples{
# see ?transfer_entropy
}
