% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEclass.R
\name{summary.transfer_entropy}
\alias{summary.transfer_entropy}
\title{Prints a summary of a transfer-entropy result}
\usage{
\method{summary}{transfer_entropy}(object, ...)
}
\arguments{
\item{object}{a transfer_entropy}

\item{...}{additional arguments, currently not in use}
}
\value{
invisible the text
}
\description{
Prints a summary of a transfer-entropy result
}
\examples{
# see ?transfer_entropy
}
