% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TE.R
\name{TE}
\alias{TE}
\title{Wrapper for the implementation of Shannon and Renyi transfer entropy.
--------------------------PRELIMINARY-------------------------------}
\usage{
TE(x, lx = 1, y, ly = 1, q = 0.5, entropy = "Shannon", shuffle = TRUE,
  const = FALSE, constx = 0, consty = 0, nreps = 2, shuffles = 6,
  ncores = parallel::detectCores() - 1, type = "quantiles",
  quantiles = c(5, 95), bins = NULL, limits = NULL, nboot = 1000,
  burn = 50)
}
\arguments{
\item{burn}{}
}
\value{
returns a list containing the respective transfer entropy measure,
the effective transfer entropy measure, standard errors, p-values and
number of observations
}
\description{
Wrapper for the implementation of Shannon and Renyi transfer entropy.
--------------------------PRELIMINARY-------------------------------
}
\examples{

}
