% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/te_renyi.R
\name{te_renyi}
\alias{te_renyi}
\title{Function to implement Renyi transfer entropy.}
\usage{
te_renyi(x, lx, y, ly, q, const = FALSE, constx = 0, consty = 0,
  nreps = 2, shuffles = 6, cl = NULL, type = "quantiles",
  quantiles = c(5, 95), bins = NULL, limits = NULL, nboot = 3,
  burn = 50, quiet = FALSE)
}
\arguments{
\item{x}{a vector of values}

\item{lx}{Markov order of x, defaults to 1}

\item{y}{a vector of values}

\item{ly}{Markov order of y, defaults to 1}

\item{q}{weighting parameter in Renyi transfer entropy, defaults to 0.5}

\item{constx}{constant value substracted from transfer entropy measure x,
default to NULL (no const)}

\item{consty}{constant value substracted from transfer entropy measure y,
default to NULL (no const)}

\item{nreps}{number of replications for each shuffle}

\item{shuffles}{number of shuffles}

\item{cl}{a numeric value (defaults to number of cores - 1),
or a cluster as created by \code{\link[parallel]{makeCluster}}
that can be used by \code{\link[pbapply]{pbapply}}}

\item{type}{bins, limits or quantiles of empirical distribution to discretize
the data}

\item{quantiles}{quantiles to use for discretization}

\item{bins}{the number of bins with equal width used for discretization}

\item{limits}{limits used for discretization}

\item{nboot}{number of bootstrap replications}

\item{burn}{number of observations that are dropped from the beginning of
the bootstrapped Markov chain}

\item{quiet}{if FALSE (default), the function gives feedback}
}
\value{
returns a list
}
\description{
Function to implement Renyi transfer entropy.
}
\examples{

}
\keyword{internal}
