% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/te_shannon.R
\name{te_shannon}
\alias{te_shannon}
\title{Function to implement Shannon transfer entropy.}
\usage{
te_shannon(x, lx, y, ly, shuffle = TRUE, const = FALSE, constx = 0,
  consty = 0, nreps = 2, shuffles = 6, ncores = parallel::detectCores()
  - 1, type = "quantiles", quantiles = c(5, 95), bins = NULL,
  limits = NULL, nboot, burn = 50)
}
\arguments{
\item{x}{a vector of values}

\item{lx}{Markov order of x, defaults to 1}

\item{y}{a vector of values}

\item{ly}{Markov order of y, defaults to 1}

\item{shuffle}{if TRUE (default), shuffled transfer entropy is calculated}

\item{constx}{constant value substracted from transfer entropy measure x,
default to NULL (no const)}

\item{consty}{constant value substracted from transfer entropy measure y,
default to NULL (no const)}

\item{nreps}{number of replications for each shuffle}

\item{shuffles}{number of shuffles}

\item{ncores}{number of cores in parallel computation}

\item{type}{bins, limits or quantiles of empirical distribution to discretize
the data}

\item{quantiles}{quantiles to use for discretization}

\item{bins}{the number of bins with equal width used for discretization}

\item{limits}{limits used for discretization}

\item{nboot}{number of bootstrap replications}

\item{burn}{number of observations that are dropped from the beginning of
the bootstrapped Markov chain}
}
\value{
returns a list
}
\description{
Function to implement Shannon transfer entropy.
}
\examples{

}
\keyword{internal}
