% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_boot_H0.R
\name{trans_boot_H0}
\alias{trans_boot_H0}
\title{Function for bootstrapping transfer entropy under H0.}
\usage{
trans_boot_H0(x, y, shuffle = TRUE, lx, ly, const = FALSE, constx = NULL,
  consty = NULL, nrep = 10, shuffles = 6)
}
\arguments{
\item{x}{a vector of coded values}

\item{shuffle}{if TRUE, shuffled transfer entropy is calculated}

\item{lx}{x(k)}

\item{ly}{y(j)}

\item{const}{if TRUE, then shuffle is constant for all bootstraps}

\item{shuffles}{}

\item{valuestab}{table of codes and associated values}
}
\value{
returns a vector
}
\description{
Function for bootstrapping transfer entropy under H0.
}
\examples{

}
