% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer_entropy.R
\name{transfer_entropy}
\alias{transfer_entropy}
\title{Function that calculates the transfer entropy between two time series x and
y. The information flow from y to x is measured. Change x and y in function
call to infer the dominant direction of the information flow.}
\usage{
transfer_entropy(x, y, lx, ly)
}
\arguments{
\item{x}{a vector of coded values}

\item{y}{a vector of coded values}

\item{lx}{x(k)}

\item{ly}{y(j)}
}
\value{
returns a list
}
\description{
Function that calculates the transfer entropy between two time series x and
y. The information flow from y to x is measured. Change x and y in function
call to infer the dominant direction of the information flow.
}
\examples{

}
