% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer_entropy_ren.R
\name{transfer_entropy_ren}
\alias{transfer_entropy_ren}
\title{Function that calculates the transfer entropy between two time series x and
y. The information flow from y to x is measured. Change x and y in function
call to infer the dominant direction of the information flow. Calculated
transfer entropy measure is Renyi transfer entropy.}
\usage{
transfer_entropy_ren(x, lx, y, ly, q)
}
\arguments{
\item{x}{a vector of coded values}

\item{lx}{Markov order of x}

\item{y}{a vector of coded values}

\item{ly}{Markov order of y}

\item{q}{weighting parameter in R?nyi transfer entropy}
}
\value{
returns a list
}
\description{
Function that calculates the transfer entropy between two time series x and
y. The information flow from y to x is measured. Change x and y in function
call to infer the dominant direction of the information flow. Calculated
transfer entropy measure is Renyi transfer entropy.
}
\examples{

}
\keyword{internal}
