% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_sample.R
\name{code_sample}
\alias{code_sample}
\title{Codes the sample by assigning data to bins that are, for example, based on
the quantiles of the empirical distribution of the sample.}
\usage{
code_sample(x, type = "quantiles", quantiles = c(5, 95), bins = NULL,
  limits = NULL, scale = 1e+10)
}
\arguments{
\item{x}{a vector of numerical values}

\item{type}{bins, limits or quantiles of empirical distribution to discretize
the data}

\item{quantiles}{quantiles to use for discretization}

\item{bins}{the number of bins with equal width used for discretization}

\item{limits}{limits used for discretization}

\item{scale}{a scale parameter}
}
\value{
returns a numerical vector
}
\description{
Codes the sample by assigning data to bins that are, for example, based on
the quantiles of the empirical distribution of the sample.
}
\examples{
\dontrun{
 set.seed(42)
 x <- rnorm(100)
 code_sample(x)

 # or coming from a data.table framework:
 library(data.table)
 set.seed(42)
 dt <- data.table(x = rnorm(100))
 dt[, sample := code_sample(dt$x)]
}
}
\keyword{internal}
