% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Markov_boot_step.R
\name{Markov_Boot_step}
\alias{Markov_Boot_step}
\title{Function for step-wise bootstraping the Markov chain.}
\usage{
Markov_Boot_step(x, lx, burn = 50)
}
\arguments{
\item{x}{a vector of coded values}

\item{lx}{x(k)}

\item{burn}{number of observations that are dropped from the beginning of
the bootstrapped time series}
}
\value{
returns a vector of strings
}
\description{
Function for step-wise bootstraping the Markov chain.
}
\examples{

}
