% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEclass.R
\name{print.TEResult}
\alias{print.TEResult}
\title{Prints a transfer-entropy result}
\usage{
\method{print}{TEResult}(x, digits = 4, boot = TRUE, ...)
}
\arguments{
\item{x}{a TEResult}

\item{digits}{the number of digits to display, defaults to 4}

\item{boot}{if the bootstrapped results should be printed, defaults to TRUE}

\item{...}{additional arguments, currently not in use}
}
\value{
invisible the text
}
\description{
Prints a transfer-entropy result
}
\examples{
# see ?transfer_entropy()
}
