% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffle_shannon.R
\name{shuffle_shannon}
\alias{shuffle_shannon}
\title{Function to calculate the effective transfer entropy, as given by the
difference between the Shannon transfer entropy calculated from a sample and
the respective shuffled transfer entropy.}
\usage{
shuffle_shannon(x, lx, y, ly, nreps = 2, shuffles = 6, diff = TRUE,
  ncores = parallel::detectCores() - 1)
}
\arguments{
\item{x}{a vector of coded values}

\item{lx}{Markov order of x}

\item{y}{a vector of coded values}

\item{ly}{Markov order of y}

\item{nreps}{number of replications for each shuffle}

\item{shuffles}{number of shuffles}

\item{diff}{if TRUE, the effective transfer entropy is calculated, otherwise
only the shuffled transfer entropy is returned}

\item{ncores}{number of cores in parallel computation}
}
\value{
returns a numeric scalar
}
\description{
Function to calculate the effective transfer entropy, as given by the
difference between the Shannon transfer entropy calculated from a sample and
the respective shuffled transfer entropy.
}
\examples{

}
