% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/te_class.R
\name{summary.transfer_entropy}
\alias{summary.transfer_entropy}
\title{Prints a summary of a transfer-entropy result}
\usage{
\method{summary}{transfer_entropy}(object, digits = 4, ...)
}
\arguments{
\item{object}{a transfer_entropy}

\item{digits}{the number of digits to display, defaults to 4}

\item{...}{additional arguments, passed to \code{\link[stats]{printCoefmat}}}
}
\value{
invisible the object
}
\description{
Prints a summary of a transfer-entropy result
}
\examples{
# see ?transfer_entropy
}
